Refactor command line argument handling in sema-harvest
This commit refactors the command line argument handling in the sema-harvest module. It replaces the existing argparse.ArgumentParser with the SemaArgsParser from sema.commons.cli. The logconf argument has been removed as it is now handled by the SemaArgsParser. The description of the parser has been updated to provide a more accurate description of the harvesting service. The changes also include removing the enable_logging function and updating the _main function to use the new argument parser. Finally, the commit adds a HarvestResult class to handle the result of the harvest service.

Related to: #126