fix #153 by including config folder path
the argument of root path was not given in the sema.harvest implementation.  This self.config_files_fodler variabl was added again so this path is now present again.

for the testing 2 were written to test a "correct" folder for which the _files_folder() would find files ending with *yml | *yaml and one which would result in an invalid path with no files found.